style(validation): rename the interface
- 이름 검증 인터페이스명 변경
